# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Enicilion Platform API
  description: Complete OpenAPI specification for Enicilion Enterprise Java Spring Boot Backend.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local Development Server
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user or upgrade a guest account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullName
                - email
                - password
              properties:
                fullName:
                  type: string
                  example: Jane Doe
                email:
                  type: string
                  format: email
                  example: jane@example.com
                password:
                  type: string
                  example: Password123
                whatsapp:
                  type: string
                  example: "+919876543210"
                referredBy:
                  type: string
                  example: REFC12
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or email already exists
  /api/auth/login:
    post:
      summary: Authenticate user and return access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: jane@example.com
                password:
                  type: string
                  example: Password123
      responses:
        '200':
          description: Successfully authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refresh_token=jwt_cookie; HttpOnly; Secure; Path=/; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials or account banned
  /api/auth/refresh:
    post:
      summary: Generate new access token using refresh cookie
      tags:
        - Authentication
      parameters:
        - in: cookie
          name: refresh_token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Access token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        example: new_jwt_access_token_here
        '401':
          description: Invalid or expired refresh token
...
  /api/applications:
    post:
      summary: Submit drifter showcase application with modified car specs and photos
      tags:
        - Drifter Applications
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - eventId
                - make
                - model
                - year
              properties:
                eventId:
                  type: string
                  format: uuid
                make:
                  type: string
                  example: Nissan
                model:
                  type: string
                  example: Silvia S15
                year:
                  type: integer
                  example: 2002
                color:
                  type: string
                  example: Pearl White
                licensePlate:
                  type: string
                  example: DL-3C-1234
                vin:
                  type: string
                  example: JH4DC23100Sxxxxxx
                engineSpec:
                  type: string
                  example: SR20DET
                modifications:
                  type: string
                  example: Garrett Turbo, HKS Coilover, Bride Seats
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Application successfully submitted
    get:
      summary: Fetch list of all drifter applications (Admin / Staff Only)
      tags:
        - Drifter Applications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of applications
  /api/applications/{id}/review:
    post:
      summary: Approve or reject a drifter application (Admin / Staff Only)
      tags:
        - Drifter Applications
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [approved, rejected, waitlisted]
                adminNotes:
                  type: string
                  example: Excellent vehicle specs, approved.
...
  /api/influencer/apply:
    post:
      summary: User submits application to become an affiliate influencer
      tags:
        - Influencers
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fullName
                - email
                - phone
                - socialLinks
                - followerCount
                - nicheDescription
                - paymentDetails
              properties:
                fullName:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                socialLinks:
                  type: string
                followerCount:
                  type: integer
                nicheDescription:
                  type: string
                paymentDetails:
                  type: string
      responses:
        '200':
          description: Affiliate application processed
  /api/influencer/dashboard:
    get:
      summary: Fetch influencer affiliate metrics, earnings, and coupons
      tags:
        - Influencers
      security:
        - BearerAuth: []
      responses:
        '200':
          description: dashboard details
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            access_token:
              type: string
            user:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                fullName:
                  type: string
                email:
                  type: string
                role:
                  type: string
                whatsapp:
                  type: string